requirements.txt
aiohttp==3.12.8
discord.py==2.5.2
gspread==6.2.1
oauth2client==4.1.3
